统一浮点数精度、增加微表面模型

+++ Common/Light.h
@@ -5,5 +5,5 @@
 struct Light {
     virtual void SampleHitResult(HitResult &result) = 0;
 
-    virtual double PDF() = 0;
+    virtual float PDF() = 0;
 };
