+++ Math/Random.h
@@ -6,12 +6,12 @@
 //-------------------------------------------------------------
 static std::random_device rd;
 static std::mt19937 mersenne_twister(rd());
-static std::uniform_real_distribution<double> uniform(0,1);
+static std::uniform_real_distribution<float> uniform(0,1);
 
-inline double RandomDouble() {
+inline float RandomFloat() {
     return uniform(mersenne_twister);
 }
 
-inline double RandomDouble(double min, double max) {
-    return min + (max-min) * RandomDouble();
+inline float RandomFloat(float min, float max) {
+    return min + (max-min) * RandomFloat();
 }
